+++ samples/scaleout/senderside/Version_6/ScaleOut.sln.DotSettings
