+++ samples/scaleout/senderside/Version_6/Server1/DoSomethingHandler.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+using NServiceBus;
+
+#region Server-Handler
+public class DoSomethingHandler : IHandleMessages<DoSomething>
+{
+    public Task Handle(DoSomething message, IMessageHandlerContext context)
+    {
+        Console.WriteLine("Message received.");
+        return Task.FromResult(0);
+    }
+}
+#endregion
