+++ samples/scaleout/senderside/Version_6/Server2/DoSomethingHandler.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading.Tasks;
+using NServiceBus;
+
+public class DoSomethingHandler : IHandleMessages<DoSomething>
+{
+    public Task Handle(DoSomething message, IMessageHandlerContext context)
+    {
+        Console.WriteLine("Message received.");
+        return Task.FromResult(0);
+    }
+}
