+++ samples/scaleout/senderside/Version_6/Server2/Program.cs
@@ -1,41 +1,25 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Configuration;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using Messages;
 using NServiceBus;
 
-namespace Server2
+class Program
 {
-    class Program
+    static void Main()
     {
-        static void Main(string[] args)
-        {
-            BusConfiguration busConfig = new BusConfiguration();
-            busConfig.EndpointName("Server");
-            busConfig.EndpointInstanceId(() => ConfigurationManager.AppSettings["InstanceId"]);
-            busConfig.UsePersistence<InMemoryPersistence>();
-            busConfig.SendFailedMessagesTo("error");
-            Run(busConfig).GetAwaiter().GetResult();
-        }
-
-        private static async Task Run(BusConfiguration busConfig)
-        {
-            IEndpointInstance endpoint = await Endpoint.Start(busConfig);
-            Console.WriteLine("Press <enter> to exit.");
-            Console.ReadLine();
-            await endpoint.Stop();
-        }
+        BusConfiguration busConfig = new BusConfiguration();
+        busConfig.EndpointName("Server");
+        busConfig.EndpointInstanceId(() => ConfigurationManager.AppSettings["InstanceId"]);
+        busConfig.UsePersistence<InMemoryPersistence>();
+        busConfig.SendFailedMessagesTo("error");
+        Run(busConfig).GetAwaiter().GetResult();
     }
 
-    public class DoSomethingHandler : IHandleMessages<DoSomething>
+    static async Task Run(BusConfiguration busConfig)
     {
-        public Task Handle(DoSomething message, IMessageHandlerContext context)
-        {
-            Console.WriteLine("Message received.");
-            return Task.FromResult(0);
-        }
+        IEndpointInstance endpoint = await Endpoint.Start(busConfig);
+        Console.WriteLine("Press <enter> to exit.");
+        Console.ReadLine();
+        await endpoint.Stop();
     }
 }
